Do not force sampling period and method
When passing the sampling period and method via the API,
the values given overwrite what is set in the site policies.py.
We don't want that by default.

It's still possible to pass the samplingperiod and samplingmethod to the
archive_pv (use with care).
But get_pvs_from_files only return PV names. It disregards the period
and method set in the files. Method and period are chosen by the
policies.py script.

JIRA INFRA-528