Remove dataRetrievalURL transformation
The dataRetrievalURL returned by /getApplianceInfo was transformed due
to some misunderstanding about the appliances.xml
There is a strict check on cluster_inetport and there can be issues
using the FQDN there (we now use the IP).
But we can use what we want for data_retrieval_url. We could use the
FQDN or IP.
data_retrieval_url can actually be set to a load balancer url so we
shouldn't manipulate it.